v5/requirements.txt
# requirements.txt — V2 Python Dependencies
#
# Install with:  pip install -r requirements.txt

# ── Core ML ──────────────────────────────────────────────────
torch>=2.0.0
xgboost>=1.7.0
scikit-learn>=1.3.0
scipy>=1.9.0

# ── Data ─────────────────────────────────────────────────────
numpy>=1.24.0
pandas>=2.0.0
yfinance>=0.2.28

# ── Visualisation ─────────────────────────────────────────────
matplotlib>=3.7.0

# ── Hyperparameter optimisation ───────────────────────────────
optuna>=3.0.0

# ── SHAP Explainability (NEW in V2) ──────────────────────────
shap>=0.44.0

# ── Config ────────────────────────────────────────────────────
PyYAML>=6.0

# ── System monitoring ────────────────────────────────────────
psutil>=5.9.0
